fixing type error

+++ lib/src/features/weather/data/models/main_weather_model.dart
@@ -21,8 +21,8 @@ class MainWeatherModel extends MainWeather {
     return MainWeatherModel(
       temp: json['temp'],
       feelsLike: double.parse(json['feels_like'].toString()),
-      tempMin: json['temp_min'],
-      tempMax: json['temp_max'],
+      tempMin: json['temp_min'].toDouble(),
+      tempMax: json['temp_max'].toDouble(),
       pressure: json['pressure'],
       humidity: json['humidity'],
     );
